move `main` entrypoint into ./lib
this allows us to require only parts of the library by doing a call like
`require('bencode/encode')`
