+++ lib/index.js
@@ -1,7 +1,7 @@
 var bencode = module.exports
 
-bencode.encode = require( './lib/encode' )
-bencode.decode = require( './lib/decode' )
+bencode.encode = require( './encode' )
+bencode.decode = require( './decode' )
 
 /**
  * Determines the amount of bytes
