--- test/lib.js
@@ -1,3 +1,0 @@
-module.exports = require('../bencode.js');
-//module.exports.decode = module.exports.bdecode
-//module.exports.encode = module.exports.bencode
